requirements.txt
flask==1.0.2
flask_restful==0.3.6
psycopg2-binary
gunicorn
